--- src/components/beagle-future-list-view/beagle-list-view.component.less
@@ -1,43 +1,0 @@
-/*
-  * Copyright 2020 ZUP IT SERVICOS EM TECNOLOGIA E INOVACAO SA
-  *
-  * Licensed under the Apache License, Version 2.0 (the "License");
-  * you may not use this file except in compliance with the License.
-  * You may obtain a copy of the License at
-  *
-  *  http://www.apache.org/licenses/LICENSE-2.0
-  *
-  * Unless required by applicable law or agreed to in writing, software
-  * distributed under the License is distributed on an "AS IS" BASIS,
-  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  * See the License for the specific language governing permissions and
-  * limitations under the License.
-*/
-
-beagle-future-list-view {
-  display: block;
-
-  .beagle-list-view {
-    display: flex;
-
-    &.VERTICAL {
-      flex-direction: column;
-    }
-
-    &.HORIZONTAL {
-      flex-direction: row;
-    }
-  }
-
-  &.hasScroll {
-    overflow: auto;
-
-    &.HORIZONTAL {
-      width: 100%;
-    }
-
-    &.VERTICAL {
-      height: 100%;
-    }
-  }
-}
