+++ src/components/beagle-list-view/beagle-list-view.on-init.ts
