+++ src/components/beagle-list-view/beagle-list-view.scroll.ts
@@ -89,21 +89,27 @@ export class BeagleListViewScroll
 
   canScrollContent (element: HTMLElement) {
     return (
-     this.direction === 'HORIZONTAL'
-       ? element.scrollWidth > element.clientWidth
-       : element.scrollHeight > element.clientHeight
-     )
-   }
+      this.direction === 'HORIZONTAL'
+        ? element.scrollWidth > element.clientWidth
+        : element.scrollHeight > element.clientHeight
+      )
+  }
+
+  getNodeToListenTo() {
+    return this.parentNode.nodeName === 'HTML' ? window : this.parentNode
+  }
 
   createScrollListener() {
     if (this.scrollSubscription) this.scrollSubscription.unsubscribe()
-    const listenTo = this.parentNode.nodeName === 'HTML' ? window : this.parentNode
+    const listenTo = this.getNodeToListenTo()
+    this.scrollSubscription = fromEvent(listenTo, 'scroll').subscribe(() => this.calcPercentage())
+  }
 
+  runOnScrollEndIfNotScrollable() {
+    const listenTo = this.getNodeToListenTo()
     if (!this.canScrollContent(listenTo instanceof Window ? document.body : listenTo)) {
       this.callOnScrollEnd()
     }
-    
-    this.scrollSubscription = fromEvent(listenTo, 'scroll').subscribe(() => this.calcPercentage())
   }
 
   calcPercentage() {
