+++ src/components/index.ts
@@ -16,7 +16,6 @@
 
 export * from './beagle-button/beagle-button.component'
 export * from './beagle-list-view/beagle-list-view.component'
-export * from './beagle-future-list-view/beagle-list-view.component'
 export * from './beagle-text/beagle-text.component'
 export * from './beagle-container/beagle-container.component'
 export * from './beagle-page-view/beagle-page-view.component'
