+++ src/constants.ts
@@ -18,7 +18,6 @@ import {
   BeagleButtonComponent,
   BeagleTextComponent,
   BeagleListViewComponent,
-  BeagleFutureListViewComponent,
   BeagleContainerComponent,
   BeaglePageViewComponent,
   BeagleImageComponent,
@@ -51,7 +50,6 @@ const beagleDefaultComponents = {
   'beagle:button': BeagleButtonComponent,
   'beagle:text': BeagleTextComponent,
   'beagle:listview': BeagleListViewComponent,
-  'beagle:futurelistview': BeagleFutureListViewComponent,
   'beagle:container': BeagleContainerComponent,
   'beagle:screencomponent': BeagleScreenComponent,
   'beagle:pageview': BeaglePageViewComponent,
